pyproject.toml
[tool.poetry]
name = "sdcard"
version = "0.1.0"
description = "package for managing SD cards"
authors = ["Nick Mortimer <nick.mortimer@csiro.au>"]
readme = "README.md"

[tool.poetry.scripts]
sdcard = "sdcard.main:sdcard"


[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.14.0"}
click = ">=8.0,<8.2"
pandas = "^2.2.0"
psutil = "^5.9.8"
pyyaml = "^6.0"
jinja2 = "^3.1.0"
zstandard = "^0.23.0"
wmi = {version = "^1.5.1", markers = "sys_platform == 'win32'"}
pywinusb = { version = "^0.4.2", markers = "sys_platform == 'win32'" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
